refactor: cleanup plugins
added a deletes options in open-next plugin
