--- packages/open-next/src/adapters/plugins/13.5/util.ts
@@ -1,5 +1,0 @@
-import { debug } from "../../logger.js";
-
-//#override requireHooks
-debug("No need to override require hooks with next 13.4.20+");
-//#endOverride
