--- packages/open-next/src/adapters/plugins/util.replacement.ts
@@ -1,25 +1,0 @@
-import { NextConfig } from "config/index";
-
-//#override requestHandler
-// @ts-ignore
-export const requestHandler = new NextServer.default({
-  conf: {
-    ...NextConfig,
-    // Next.js compression should be disabled because of a bug in the bundled
-    // `compression` package — https://github.com/vercel/next.js/issues/11669
-    compress: false,
-    // By default, Next.js uses local disk to store ISR cache. We will use
-    // our own cache handler to store the cache on S3.
-    experimental: {
-      ...NextConfig.experimental,
-      // This uses the request.headers.host as the URL
-      // https://github.com/vercel/next.js/blob/canary/packages/next/src/server/next-server.ts#L1749-L1754
-      trustHostHeader: true,
-      incrementalCacheHandlerPath: `${process.env.LAMBDA_TASK_ROOT}/cache.cjs`,
-    },
-  },
-  customServer: false,
-  dev: false,
-  dir: __dirname,
-}).getRequestHandler();
-//#endOverride
