+++ packages/open-next/src/core/requestHandler.ts
@@ -1,4 +1,3 @@
-import { BuildId } from "config/index";
 import {
   IncomingMessage,
   OpenNextNodeResponse,
@@ -7,10 +6,9 @@ import {
 import { InternalEvent, InternalResult } from "types/open-next";
 
 import { error } from "../adapters/logger";
-import { createServerResponse } from "../adapters/plugins/routing/util";
-import { handler as serverHandler } from "../adapters/plugins/serverHandler";
-import { convertRes } from "./routing/util";
+import { convertRes, createServerResponse, proxyRequest } from "./routing/util";
 import routingHandler from "./routingHandler";
+import { requestHandler, setNextjsPrebundledReact } from "./util";
 
 export async function openNextHandler(
   internalEvent: InternalEvent,
@@ -20,8 +18,9 @@ export async function openNextHandler(
     internalEvent.headers.host = internalEvent.headers["x-forwarded-host"];
   }
 
-  //TODO: replace this line for next <= 13.4.12
+  //#override withRouting
   const preprocessResult = await routingHandler(internalEvent);
+  //#endOverride
 
   if ("type" in preprocessResult) {
     // res is used only in the streaming case
@@ -88,11 +87,19 @@ async function processRequest(
   try {
     // `serverHandler` is replaced at build time depending on user's
     // nextjs version to patch Nextjs 13.4.x and future breaking changes.
-    await serverHandler(req, res, {
-      internalEvent,
-      buildId: BuildId,
-      isExternalRewrite,
-    });
+
+    const { rawPath } = internalEvent;
+
+    if (isExternalRewrite) {
+      return proxyRequest(req, res);
+    } else {
+      //#override applyNextjsPrebundledReact
+      setNextjsPrebundledReact(rawPath);
+      //#endOverride
+
+      // Next Server
+      await requestHandler(req, res);
+    }
   } catch (e: any) {
     error("NextJS request failed.", e);
     //TODO: we could return the next 500 page here
